Move Flatten to static class
